Remove aad to encrypt all data in config
Receiving InvalidTag error when encrypting all or most entries in the config.
This is thrown when the ciphertext has been changed or the aad is invalid.
Since the error could not be narrowed down this meant the removal of aad.
This means cloudmesh will not authenticate additional data when decrypting.
The cipher already has integrity gurantees but the extra protection of checking
the version number will be lost. This is a minor loss for software stability.